finished openssh packets and started fs implementation

+++ sshd-fs/src/main/java/org/apache/sftp/protocol/impl/MaskFactory.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 
 public class MaskFactory {
-    public static <E extends Enum<E> & Maskable<E>> EnumSet<E> fromMask( int mask, Class<E> type ) {
+    public static <E extends Enum<E> & Maskable<E>> EnumSet<E> fromMask( long mask, Class<E> type ) {
         List<E> maskList = new ArrayList<>();
         for ( E maskValue : type.getEnumConstants() ) {
             if ( (mask & maskValue.getValue()) == maskValue.getValue() ) {
@@ -17,7 +17,7 @@ public class MaskFactory {
         return maskList.isEmpty() ? EnumSet.noneOf( type ) : EnumSet.copyOf( maskList );
     }
 
-    public static <E extends Enum<E> & Maskable<E>> int toMask( EnumSet<E> enumSet ) {
+    public static <E extends Enum<E> & Maskable<E>> long toMask( EnumSet<E> enumSet ) {
         int mask = 0;
         if ( enumSet != null && !enumSet.isEmpty() ) {
             for ( E maskValue : enumSet ) {
